Add files via upload
This is a monolithic import of a RAD example I had permission to publish. It's a working demo that was soon afterwards renamed and brought into the fold behind the corporate firewall and policies. Test data that was more extensive and based on real data was later created and randomised to protect individuals' privacy. It's based on a Flask quickstart example of a ToDo application that worked "out of the box" for me better than the Pyramid one I initially tried.